+++ ChangeLog.md
@@ -2,9 +2,9 @@
 
 ## ChangeLog
 
-### 1.2.0-rc
+### 1.2.0
 
-    Release date: 
+    Release date: 2023-10-11
 
     - Define new Roles.countByRole() publication (bumping candidate version number)
     - Reorganize constants definitions to not pollute global space
@@ -56,4 +56,4 @@
 
 ---
 P. Wieser
-- Last updated on 2023, June 20th
+- Last updated on 2023, Oct. 11th